+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Unreleased
 
 .. No unreleased changes.
 
+* Added a player-visible Cursor tool.
 * Added icon to the "About DMap" menu item in the Help menu.
 * Updated the “Reset View” action: renamed to “Refit View”, updated icon, and now fits the map to the window.
 * Linux release builds are now single-file executables.